🚚 Divide Project+Templates into separate files.
